adjust min/max entropy picking values
A hobbyist password cracking machine can work through every unsalted SHA-1
digest built from 8-character passwords in one week. If the password is
randomly and uniformly selected, then it comes from about 52-bits of entropy.
However, if the password were 9 characters in length, then it would come from
about 58-bits of entropy. This would take 94 weeks for the same hardware to
completely exhaust.

Thus, the absolute bottom-of-the-barrel entropy should be 55-bits. So, the min
and max values were adjusted to reflect this