Switch to the PGP wordlist as default for alternate
The PGP wordlist is designed to be spoken in noisy envioronments and still be
unambiguous about what was spoken. This makes the PGP word list ideal for
speaking passwords to a coworker in a noisy server room or saying it over the
phone to a software vendor providing support for a product.

So I think there is value to the community for providing this as a selected
default over the Simpsons, as memorable as some of those quotes could be.