Update and cleanup the Trump word list
My scripts parsing the @realDonaldTrump Twitter timeline were
good-but-not-great. I've since then greatly expanded them to vastly improve the
quality of the word list.

This commit brings to total up to 7,789 unique words, 403 total words from
exactly 13-bits of entropy per word. Once I reach this point, the longer words
will be removed from the word list in favor of the shorter ones, to bring the
average passphrase length down to more reasonable levels.

This will take time, of course, but the end goal is to get as close as possible
to the Diceware word length averages, while still maintaining the unique
personality Trump brings to Twitter.