Aggressively reduce word list file sizes
Instead of one comma and two quotes per word to separate the words in each word
list, each word list now is dynamically built by separating the words into
their own array based on word length. This allows me to deterministically get
each word without storing an unnecessary data.

As a demonstration:

    * Old: alternate_pgp.e0ed41a5.js, 5476 bytes
    * New: alternate_pgp.b1e5fa80.js, 4148 bytes

    * Old: alternate_rockyou.f0ffe682.js, 77897 bytes
    * New: alternate_rockyou.b0287590.js, 54879 bytes

    * Old: alternate_simpsons.075b9d4e.js, 44974 bytes
    * New: alternate_simpsons.64f6d017.js, 30274 bytes

    * Old: alternate_trump.4462c2a2.js, 74091 bytes
    * New: alternate_trump.305c5d54.js, 49727 bytes

    * Old: diceware_en.f46f1222.js, 56308 bytes
    * New: diceware_en.48869f2c.js, 33155 bytes

The old lists require 258,746 bytes of storage. The new lists require 172,183
bytes of storage. That's a storage savings of 33.45%, which is significant.

This is the first commit in many to come reducing the 2.6 MB download to an
estimated 1.6 MB, give or take when all is finished. Losing a full megabyte on
download is a huge win.