Aggressive word list size reduction
As mentioned the recent previous commits, rather that comma-separating a list
of quoted words in an array, and storing an extra 3 bytes of metadat per word,
the words are sorted by the number of characters. Each character length then
has its own array with the words of that length in it. As such, the word list
is dynamically calculated on page load.

Before this new efficient approach, the total size in bytes of JavaScript word
lists was 2680558 bytes. After removing all that quote and comma metadata, the
new word lists occupy a total of 1778424 bytes.

    2680558-1778424=902134

The result of this new storage efficiency removed nearly 1 MB of data out of
the download, or about 33.65% savings. I'll take it.

I didn't reduce the size of the "Colors" dictionary, nor the "Emoji" list, as
these would require special care. I could probably squeeze mabe another 15 KB,
but at this point, I'm willing to call it "good enough".