Fix Turkish word list
I suspect that the Turkish word list inspired the Slovenian word list, or vice
versa, as they are _very_ similar. The Turkish word list provided by Mert
Dirik, and which is only available on the Internet Archive, duplicates "def".
As such, like Slovenian, I added "2a" to bring the unique word list to 7,776.