Enforce only unique elements in returned array
I found that may of the word lists I was using had duplicates in the list, or
were not as long as initially expected:

    * Alternate Elvish
    * Diceware Bulgarian
    * Diceware Italian
    * Diceware Maori
    * Diceware Spanish
    * Diceware Swedish
    * Diceware Turkish

Using the new Set() object from ES6, I am now only returning unique elements in
the list, ensuring security. Even though this actually didn't change the
resulting passphrase length after implementing it, this will catch any future
imported lists that are thought to be a certain size, but are shortened due to
dupes.