Add a DiceKeys generator
A clean room online implementation of the DiceKeys product, build strictly by observing the physical properties of the
dice and their behavior. This will produce a key that can be quickly scanned with the online app, and hopefully mobile
apps for Android and iOS when released.

The DiceKeys product is a physical 198-bit random number generator using 25 unique dice. It's designed to be permanently
snapped into place after the dice are shaken and placed in each slot, then scanned with a mobile app to
deterministically generate and regenerate a secret, such as a master passphrase for a password manager, the seed for a
BIP-39 mnemonic, or the seed for a U2F hardware key.

The physical device is not meant to generate multiple different secrets from the dice. That's where this comes in. If
you would rather generate multiple secrets with the DiceKeys product, instructions are provided with this implementation
for some ideas on how do to so. Think of it as a Rube Goldberg machine of sorts. Rather than just using the passphrase
generator to generate a passphrase, you:

1. Generate a random DiceKey
2. Record the results
3. Convert the result to a password or passphrase

As such, this implementation of the physical device is meant to be used as for demonstration purposes, testing, or as a
talking point only.

Please purchase the physical device if this interests you.