Checksums added to Bitcoin, Bubble Babble, Base32
Bubble Babble, Base32, and Bitcoin BIP 0039 are all designed to encode raw binary into manageable
data for humans. All three have a checksum built into their specification (Crockford base32, not
other base32 specs). Checksums will not be arbitrarily added to every generator however; only those
where it is defined in the specification.

This meant re-doing the Bitcoin word lists, which means increasing the space requirements
unfortunatley. Further, the keyspace in the Bitcoin passphrases must be a multiple of 32 bits, per
the spec. So when loading the page, even though 70 bits is the default minimum security margin, the
Bitcoin passphrases will have 96 bits. This makes the Bitcoin passphrase the strongest password on a
default page load.