Pseudoword changes
Removed "Korean K-Pop" and "Secret Ninja". I added them when I first started working on this project
years ago as "filler" for the pseudowords category. But in all honesty, they've always been a thorn
in my side, and they've really never offered any value to the project.

Those were replaced with "Munemo", which is a base-100 integer-to-string encoding format, which
provides a unique feature to the password generation project. The other is "Letterblock Diceware"
which uses English bigrams weights to help shape memorable but not necessarily pronounceable
passwords. Jury is out if it's really doing what it's designed to do however.

CSS changes were made to accommodate longer options in the drop down menus.