Change Munemo algorithm. Add negative numbers.
Munemo supports negative integers with the 'xa' prefix. This adds that support, but instead of
picking elements randomly from an array to build a string, we generate a random number, then use the
specification to convert it to a string using the same array.

Then at the end, we flip a coin to decide if it's negative, and prepend 'xa' if true.