CSS work for input slider
Change the low end of the range from 56 bits to 48 bits. This means that passwords can be trivially
cracked with hobbyist GPU clusters if the password is hashed with a generic hashing function, such
as MD5, SHA-1/2/3. However, the slider track turns red to indicate serious warning.

The slider track changes colors with the following parameters:

- 48 bits: red
- 56 bits: orange
- 64 bits: yellow
- 72+ bits: green

Styling is applied for both light and dark themes.